# Benchmarks and some optimizations