* Add benchmarks
* Inline functions where benchmarks show a clear decrease in time and allocations.

<details>
<summary>Benchmarks before, GHC 9.6.3 -O</summary>

```
All
  Decode
    string:      OK
      31.7 ns ± 1.9 ns, 191 B  allocated,   0 B  copied, 9.0 MB peak memory
    integer:     OK
      46.7 ms ± 4.7 ms,  27 MB allocated, 9.7 MB copied,  20 MB peak memory
    list:        OK
      175  ms ± 4.3 ms, 251 MB allocated, 248 MB copied, 138 MB peak memory
    dict:        OK
      29.4 ms ± 1.6 ms,  55 MB allocated,  42 MB copied, 138 MB peak memory
    text:        OK
      41.0 μs ± 2.7 μs, 975 KB allocated,  90 B  copied, 138 MB peak memory
    list string: OK
      38.4 ms ± 1.5 ms,  75 MB allocated,  57 MB copied, 138 MB peak memory
    list int:    OK
      220  ms ± 7.1 ms, 335 MB allocated, 302 MB copied, 237 MB peak memory
    list list:   OK
      141  ms ± 5.0 ms, 250 MB allocated, 206 MB copied, 237 MB peak memory
    list dict:   OK
      123  ms ± 2.7 ms, 244 MB allocated, 160 MB copied, 237 MB peak memory
    list fields: OK
      189  ms ± 6.6 ms, 358 MB allocated, 164 MB copied, 237 MB peak memory
    list word16: OK
      219  ms ±  14 ms, 351 MB allocated, 296 MB copied, 250 MB peak memory
  Encode
    string:      OK
      13.8 ms ± 733 μs, 101 MB allocated,  11 KB copied, 250 MB peak memory
    integer:     OK
      83.4 ms ± 3.2 ms, 368 MB allocated,  52 KB copied, 250 MB peak memory
    list:        OK
      28.1 ms ± 2.8 ms, 142 MB allocated,  26 KB copied, 250 MB peak memory
    list fusion: OK
      99.2 ms ± 6.6 ms, 236 MB allocated,  62 MB copied, 250 MB peak memory
    dict:        OK
      36.6 ms ± 2.2 ms, 169 MB allocated,  48 KB copied, 250 MB peak memory
    text:        OK
      73.3 ms ± 4.3 ms, 321 MB allocated,  42 KB copied, 250 MB peak memory
    int:         OK
      18.2 ms ± 1.4 ms,  33 MB allocated, 3.6 KB copied, 250 MB peak memory
    field:       OK
      31.2 ms ± 1.5 ms, 146 MB allocated,  20 KB copied, 250 MB peak memory
    word16:      OK
      8.43 ms ± 788 μs,  28 MB allocated, 3.0 KB copied, 250 MB peak memory
```
</details>

<details>
<summary>After</summary>

```
All
  Decode
    string:      OK
      25.8 ns ± 1.7 ns, 191 B  allocated,   0 B  copied, 9.0 MB peak memory, 18% less than baseline
    integer:     OK
      46.8 ms ± 2.1 ms,  28 MB allocated, 9.7 MB copied,  22 MB peak memory,       same as baseline
    list:        OK
      175  ms ± 6.6 ms, 251 MB allocated, 248 MB copied, 138 MB peak memory,       same as baseline
    dict:        OK
      29.5 ms ± 680 μs,  55 MB allocated,  42 MB copied, 138 MB peak memory,       same as baseline
    text:        OK
      40.6 μs ± 1.9 μs, 976 KB allocated,  89 B  copied, 138 MB peak memory,       same as baseline
    list string: OK
      38.9 ms ± 2.4 ms,  75 MB allocated,  57 MB copied, 138 MB peak memory,       same as baseline
    list int:    OK
      127  ms ± 3.8 ms, 190 MB allocated, 172 MB copied, 177 MB peak memory, 42% less than baseline
    list list:   OK
      146  ms ±  12 ms, 250 MB allocated, 206 MB copied, 177 MB peak memory,       same as baseline
    list dict:   OK
      125  ms ± 4.9 ms, 244 MB allocated, 160 MB copied, 177 MB peak memory,       same as baseline
    list fields: OK
      192  ms ± 6.2 ms, 358 MB allocated, 178 MB copied, 241 MB peak memory,       same as baseline
    list word16: OK
      143  ms ±  13 ms, 205 MB allocated, 191 MB copied, 241 MB peak memory, 34% less than baseline
  Encode
    string:      OK
      13.5 ms ± 742 μs, 101 MB allocated,  11 KB copied, 241 MB peak memory,       same as baseline
    integer:     OK
      83.5 ms ± 4.7 ms, 368 MB allocated,  52 KB copied, 241 MB peak memory,       same as baseline
    list:        OK
      22.1 ms ± 1.4 ms, 123 MB allocated,  23 KB copied, 241 MB peak memory, 21% less than baseline
    list fusion: OK
      9.16 ms ± 784 μs,  30 MB allocated, 3.3 KB copied, 241 MB peak memory, 90% less than baseline
    dict:        OK
      31.3 ms ± 1.8 ms, 139 MB allocated,  35 KB copied, 241 MB peak memory, 14% less than baseline
    text:        OK
      48.9 ms ± 3.4 ms, 276 MB allocated,  37 KB copied, 241 MB peak memory, 33% less than baseline
    int:         OK
      18.0 ms ± 1.4 ms,  33 MB allocated, 3.6 KB copied, 241 MB peak memory,       same as baseline
    field:       OK
      31.2 ms ± 1.6 ms, 143 MB allocated,  20 KB copied, 241 MB peak memory,       same as baseline
    word16:      OK
      8.30 ms ± 705 μs,  28 MB allocated, 2.9 KB copied, 241 MB peak memory,       same as baseline
```
</details>